# Optimization for long prompt