--- a/llama/generation.py
+++ b/llama/generation.py
@@ -43,20 +43,6 @@ class LLaMA:
 
         return tokens, input_tokens, cur_pos_tensor, input_pos_tensor, output_pos_tensor, cache_kvs
 
-#    def _create_start_pos_buckets(self, max_seq_len):
-#        buckets = [max_seq_len]
-#        while buckets[-1] > 64:
-#            buckets.append(buckets[-1] // 2)
-#        buckets.append(1)
-#        buckets.reverse()
-#
-#        return buckets
-#
-#    def _select_start_pos_bucket(self, prompt_size, buckets):
-#        for i in range(1, len(buckets)):
-#            if prompt_size < buckets[i]:
-#                return buckets[i - 1]
-
     def generate(
         self,
         prompts: List[str],
@@ -85,14 +71,6 @@ class LLaMA:
         tokens = tokens.to(device)
         input_text_mask = tokens != self.tokenizer.pad_id
 
-        # start_pos = 1
-        # start_pos_buckets = self._create_start_pos_buckets(params.max_seq_len)
-        # start_pos = self._select_start_pos_bucket(min_prompt_size, start_pos_buckets)
-        # print(f"start_pos = {start_pos}")
-        # cur_pos_tensor = torch.tensor(start_pos).to(device)
-        # input_pos_tensor = torch.arange(0, start_pos).to(device)
-        # output_pos_tensor = cur_pos_tensor - 1
-        # input_tokens = tokens.index_select(1, input_pos_tensor)
         cache_kvs = self.model.cache_kvs
         xm.mark_step()