@AlexWertheim 2023-05-12
LGTM, really clever approach!